+++ CHANGELOG.md
@@ -24,6 +24,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - No longer fail on relative links to directories
 - No longer fail on broken links
 - Properly format headings with embedded markdown
+- Handle constructing anchor links for repeated headings
 
 ---